Add: SLAC handler in slac_session.py Change: integration of SLAC in Change: rename secc-main
